Create standalone Validator object
